refactor(auth): consolidate auth pages into @getmunin/dashboard-pages
Promote LoginForm, SignupForm, ForgotPasswordPage, ResetPasswordPage,
VerifyEmailPage, and AuthLoading into the shared package so OSS and cloud
can share one implementation. Each accepts the brand footer as a prop;
SignupForm keeps OSS's invite-token lookup and gains OAuth provider
buttons; LoginForm moves the forgot-password link into the footnote
("New here? Create an account · Forgot password?") and shortens the
label from "Forgot your password?" to "Forgot password?". The shared
useTranslateError is now re-exported from the package root, replacing
the apps/web duplicate.
