+++ packages/dashboard-pages/src/components/auth-shell/auth-loading.tsx
@@ -1,0 +1,5 @@
+import { PageSpinner } from '@getmunin/ui';
+
+export function AuthLoading() {
+  return <PageSpinner className="min-h-screen" />;
+}
